+++ src/SmugMugModel.v2/Types/Entities/AlbumImage/AlbumImage.properties.cs
@@ -8,7 +8,7 @@ namespace SmugMug.v2.Types
 {
     public partial class AlbumImageEntity : SmugMugEntity
     {
-        private int _altitude;
+        private long _altitude;
         private bool _canEdit;
         private string _caption;
         private bool _collectable;
@@ -26,9 +26,9 @@ namespace SmugMug.v2.Types
         private string _latitude;
         private string _longitude;
         private bool _movable;
-        private int _originalHeight;
-        private int _originalSize;
-        private int _originalWidth;
+        private long _originalHeight;
+        private long _originalSize;
+        private long _originalWidth;
         private bool _processing;
         private bool _protected;
         private string _thumbnailUrl;
@@ -41,7 +41,7 @@ namespace SmugMug.v2.Types
         private string _webUri;
 
 
-        public int Altitude {
+        public long Altitude {
             get {
                 return _altitude;
             }
@@ -293,7 +293,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int OriginalHeight {
+        public long OriginalHeight {
             get {
                 return _originalHeight;
             }
@@ -307,7 +307,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int OriginalSize {
+        public long OriginalSize {
             get {
                 return _originalSize;
             }
@@ -321,7 +321,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int OriginalWidth {
+        public long OriginalWidth {
             get {
                 return _originalWidth;
             }
