+++ src/SmugMugModel.v2/Types/Entities/Download/Download.properties.cs
@@ -11,7 +11,7 @@ namespace SmugMug.v2.Types
         private DateTime _dateAdded;
         private DateTime _dateModified;
         private string _fileName;
-        private int _part;
+        private long _part;
         private ImageStatusEnum _status;
 
 
@@ -57,7 +57,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Part {
+        public long Part {
             get {
                 return _part;
             }
