+++ src/SmugMugModel.v2/Types/Entities/ImageMetadata/ImageMetadata.properties.cs
@@ -8,7 +8,7 @@ namespace SmugMug.v2.Types
 {
     public partial class ImageMetadataEntity : SmugMugEntity
     {
-        private int _altitude;
+        private long _altitude;
         private string _altitudeReference;
         private string _aperture;
         private string _audioCodec;
@@ -51,12 +51,12 @@ namespace SmugMug.v2.Types
         private string _hyperfocalDistance;
         private string _iSO;
         private string _keywords;
-        private int _latitude;
+        private long _latitude;
         private string _latitudeReference;
         private string _lens;
         private string _lensSerialNumber;
         private string _lightSource;
-        private int _longitude;
+        private long _longitude;
         private string _longitudeReference;
         private string _make;
         private string _metering;
@@ -90,7 +90,7 @@ namespace SmugMug.v2.Types
         private string _writerEditor;
 
 
-        public int Altitude {
+        public long Altitude {
             get {
                 return _altitude;
             }
@@ -692,7 +692,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Latitude {
+        public long Latitude {
             get {
                 return _latitude;
             }
@@ -762,7 +762,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Longitude {
+        public long Longitude {
             get {
                 return _longitude;
             }
