+++ src/SmugMugModel.v2/Types/Entities/ImageSizeX3Large/ImageSizeX3Large.properties.cs
@@ -11,15 +11,15 @@ namespace SmugMug.v2.Types
         private decimal _bitrate;
         private decimal _duration;
         private string _ext;
-        private int _height;
+        private long _height;
         private string _mD5;
-        private int _size;
+        private long _size;
         private string _uri;
         private string _uriDescription;
         private string _url;
         private bool _usable;
         private bool _watermarked;
-        private int _width;
+        private long _width;
 
 
         public decimal Bitrate {
@@ -64,7 +64,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Height {
+        public long Height {
             get {
                 return _height;
             }
@@ -92,7 +92,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Size {
+        public long Size {
             get {
                 return _size;
             }
@@ -176,7 +176,7 @@ namespace SmugMug.v2.Types
         }
 
 
-        public int Width {
+        public long Width {
             get {
                 return _width;
             }
